Ring : separeted NTT tests
